# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Talk Proposals API",
        "description": "API documentation for the Talk Proposals system. This API allows speakers to submit talk proposals, reviewers to review them, and admins to manage the proposal status.",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "List all tags",
                "description": "Retrieves all available tags. Supports optional search by name.",
                "operationId": "8767fe41ed2e75df7ce6f99c8aeb47cb",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search tags by name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255,
                            "example": "Technology"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "maximum": 100,
                            "minimum": 1,
                            "example": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tags retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tags retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "tags": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Tag"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 75
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    },
                    "422": {
                        "description": "Invalid query parameters"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Create a new tag",
                "description": "Creates a new tag or returns the existing tag if a tag with the same name already exists.",
                "operationId": "386137f14d37c8d53bcb5869d04e5595",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Tag name",
                                        "type": "string",
                                        "example": "Technology"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tag created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tag created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "tag": {
                                                    "$ref": "#/components/schemas/Tag"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    },
                    "419": {
                        "description": "Missing or expired CSRF token for stateful SPA requests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Tag": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Technology"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Sanctum SPA session cookie. Fetch /sanctum/csrf-cookie before login and mutations, then send the X-XSRF-TOKEN header with cookies. The generated cookie name follows SESSION_COOKIE. Authenticate via the frontend or this API; session cookies cannot be entered manually in Swagger UI.",
                "name": "laravel-session",
                "in": "cookie"
            },
            "bearerAuth": {
                "type": "http",
                "description": "An existing Sanctum personal access token. Login/register use SPA sessions and do not issue bearer tokens.",
                "scheme": "bearer"
            }
        }
    },
...
}